+++ src/Langgas/SisdikBundle/Controller/LayananSmsPeriodikController.php
@@ -8,11 +8,12 @@ use Langgas\SisdikBundle\Entity\JadwalKehadiran;
 use Langgas\SisdikBundle\Entity\PilihanLayananSms;
 use Langgas\SisdikBundle\Entity\LayananSmsPeriodik;
 use Langgas\SisdikBundle\Entity\Sekolah;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use JMS\SecurityExtraBundle\Annotation\PreAuthorize;
 
 /**
@@ -148,6 +149,10 @@ class LayananSmsPeriodikController extends Controller
             throw $this->createNotFoundException('Entity LayananSmsPeriodik tak ditemukan.');
         }
 
+        if ($this->get('security.context')->isGranted('view', $entity) === false) {
+            throw new AccessDeniedException($this->get('translator')->trans('akses.ditolak'));
+        }
+
         $deleteForm = $this->createDeleteForm($id);
 
         return [
@@ -175,8 +180,12 @@ class LayananSmsPeriodikController extends Controller
 
         $entity = $em->getRepository('LanggasSisdikBundle:LayananSmsPeriodik')->find($id);
 
-        if (! $entity) {
+        if (!$entity) {
             throw $this->createNotFoundException('Entity LayananSmsPeriodik tak ditemukan.');
+        }
+
+        if ($this->get('security.context')->isGranted('edit', $entity) === false) {
+            throw new AccessDeniedException($this->get('translator')->trans('akses.ditolak'));
         }
 
         $editForm = $this->createForm('sisdik_layanansmsperiodik', $entity);
@@ -206,6 +215,10 @@ class LayananSmsPeriodikController extends Controller
 
         if (!$entity) {
             throw $this->createNotFoundException('Entity LayananSmsPeriodik tak ditemukan.');
+        }
+
+        if ($this->get('security.context')->isGranted('edit', $entity) === false) {
+            throw new AccessDeniedException($this->get('translator')->trans('akses.ditolak'));
         }
 
         $deleteForm = $this->createDeleteForm($id);
@@ -266,6 +279,10 @@ class LayananSmsPeriodikController extends Controller
 
             if (!$entity) {
                 throw $this->createNotFoundException('Entity LayananSmsPeriodik tak ditemukan.');
+            }
+
+            if ($this->get('security.context')->isGranted('delete', $entity) === false) {
+                throw new AccessDeniedException($this->get('translator')->trans('akses.ditolak'));
             }
 
             $em->remove($entity);
